// src/types/playground.ts
export interface ToolCall {
    role: 'user' | 'tool' | 'system' | 'assistant'
    content: string | null
    tool_call_id: string
    tool_name: string
    tool_args: Record<string, string>
    tool_call_error: boolean
    metrics: {
        time: number
    }
    created_at: number
}

export interface ReasoningSteps {
    title: string
    action?: string
    result: string
    reasoning: string
    confidence?: number
    next_action?: string
}
// ...
export interface ResponseAudio {
    id?: string
    content?: string
    transcript?: string
    channels?: number
    sample_rate?: number
}
// ...
export interface ReasoningMessage {
    role: 'user' | 'tool' | 'system' | 'assistant'
    content: string | null
    tool_call_id?: string
    tool_name?: string
    tool_args?: Record<string, string>
    tool_call_error?: boolean
    metrics?: {
        time: number
    }
    created_at?: number
}
export interface PlaygroundChatMessage {
    role: 'user' | 'agent' | 'system' | 'tool'
    content: string
    streamingError?: boolean
    created_at: number
    tool_calls?: ToolCall[]
    extra_data?: {
        reasoning_steps?: ReasoningSteps[]
        reasoning_messages?: ReasoningMessage[]
        references?: ReferenceData[]
    }
    images?: ImageData[]
    videos?: VideoData[]
    audio?: AudioData[]
    response_audio?: ResponseAudio
}

export interface ComboboxAgent {
    value: string
    label: string
    model: {
        provider: string
    }
    storage?: boolean
}
export interface ImageData {
    revised_prompt: string
    url: string
}

export interface VideoData {
    id: number
    eta: number
    url: string
}

export interface AudioData {
    base64_audio?: string
    mime_type?: string
    url?: string
    id?: string
    content?: string
    channels?: number
    sample_rate?: number
}

export interface ReferenceData {
    query: string
    references: Reference[]
    time?: number
}

export interface Reference {
    content: string
    meta_data: {
        chunk: number
        chunk_size: number
    }
    name: string
}

export interface SessionEntry {
    session_id: string
    title: string
    created_at: number
}